// temper/shared-addon-libraries-lib-data-encode/src/types.ts
export type LdeValue = string | number

export type LuaTable = Record<string | number, unknown>
// ...
export interface DecodeInstance {
  encodedStrings: string[]
  currentStringIndex: number
  currentStringPos: number
  currentString: string | undefined
  currentStringLength: number | undefined
  dictionary: LdeValue[]
  data: unknown
  Initialize: (this: DecodeInstance, encodedData: string[], globalDict?: LdeValue[]) => void
  InitDictionary: (this: DecodeInstance, globalDict?: LdeValue[]) => void
  GetCurrentString: (this: DecodeInstance) => string | undefined
  GetNextChar: (this: DecodeInstance, noPosIncrement?: boolean) => string
  GetEncodedItem: (this: DecodeInstance, length?: number) => string
  MoveCurrentPos: (this: DecodeInstance, offset: number) => void
  DecodeItem: (this: DecodeInstance) => unknown
  DecodeBool: (this: DecodeInstance, controlChar: string) => boolean | undefined
  DecodeStringId: (this: DecodeInstance, controlChar: string) => LdeValue
  DecodeBase: (this: DecodeInstance, encodedItem: string) => number
  DecodeString: (this: DecodeInstance, controlChar: string) => string
  DecodeArray: (this: DecodeInstance) => unknown[]
  DecodeTable: (this: DecodeInstance) => LuaTable
  DecodeInteger: (this: DecodeInstance) => number
  DecodeNumeric: (this: DecodeInstance) => number | undefined
  [key: string]: unknown
}

export interface DecodeClass {
  Subclass: (this: DecodeClass) => DecodeClass
  New: (this: DecodeClass, encodedData: string[], globalDict?: LdeValue[]) => DecodeInstance
  Initialize: (this: DecodeInstance, encodedData: string[], globalDict?: LdeValue[]) => void
  InitDictionary: (this: DecodeInstance, globalDict?: LdeValue[]) => void
  GetCurrentString: (this: DecodeInstance) => string | undefined
  GetNextChar: (this: DecodeInstance, noPosIncrement?: boolean) => string
  GetEncodedItem: (this: DecodeInstance, length?: number) => string
  MoveCurrentPos: (this: DecodeInstance, offset: number) => void
  DecodeItem: (this: DecodeInstance) => unknown
  DecodeBool: (this: DecodeInstance, controlChar: string) => boolean | undefined
  DecodeStringId: (this: DecodeInstance, controlChar: string) => LdeValue
  DecodeBase: (this: DecodeInstance, encodedItem: string) => number
  DecodeString: (this: DecodeInstance, controlChar: string) => string
  DecodeArray: (this: DecodeInstance) => unknown[]
  DecodeTable: (this: DecodeInstance) => LuaTable
  DecodeInteger: (this: DecodeInstance) => number
  DecodeNumeric: (this: DecodeInstance) => number | undefined
  [key: string]: unknown
}